add fetchconfig to section "Backups"
The existing entry for "Jazigo" mentions "fetchconfig," but the list
does not yet contain a link to it.

If you look at the old fetchconfig site hosted on Savannah
( http://www.nongnu.org/fetchconfig/ ) you will see that
fetchconfig is now hosted at GitHub, which is the link I add.