Correct minor issues within README.md
Correct spelling mistakes throughout document. Correct common product
acronym misspellings, such as changing "netbox" and "Netbox" to "NetBox".
Ensure products, languages, and so on are capitlized ("python" to
"Python"). Reworded the document header as well as a few resource
descriptions to be more grammatically correct and/or make more sense.